injector: simplify port exclusion code
Simplifies the code and tests related to port
exclusions and consolidates exclusions related
code in a separate file for better maintenance.
