+++ pkg/injector/webhook.go
@@ -334,28 +334,6 @@ func (wh *mutatingWebhook) mustInject(pod *corev1.Pod, namespace string) (bool, 
 	return false, nil
 }
 
-// getPortExclusionListForPod gets a list of ports to exclude from sidecar traffic interception for the given
-// pod and annotation kind.
-//
-// Ports are excluded from sidecar interception when the pod is explicitly annotated with a single or
-// comma separate list of ports.
-//
-// The kind of exclusion (inbound vs outbound) is determined by the specified annotation.
-//
-// The function returns an error when it is unable to determine whether ports need to be excluded from outbound sidecar interception.
-func (wh *mutatingWebhook) getPortExclusionListForPod(pod *corev1.Pod, namespace string, annotation string) ([]int, error) {
-	var ports []int
-	// Check if the pod is annotated for outbound port exclusion
-	ports, err := isAnnotatedForPortExclusion(pod.Annotations, annotation, pod.Kind, fmt.Sprintf("%s/%s", pod.Namespace, pod.Name))
-	if err != nil {
-		log.Error().Err(err).Str(errcode.Kind, errcode.GetErrCodeWithMetric(errcode.ErrDeterminingPodPortExclusions)).
-			Msgf("Error determining port exclusions for annotation %s on pod %s/%s", annotation, namespace, pod.Name)
-		return ports, err
-	}
-
-	return ports, nil
-}
-
 func isAnnotatedForInjection(annotations map[string]string, objectKind string, objectName string) (exists bool, enabled bool, err error) {
 	inject, ok := annotations[constants.SidecarInjectionAnnotation]
 	if !ok {
@@ -373,27 +351,6 @@ func isAnnotatedForInjection(annotations map[string]string, objectKind string, o
 		err = errors.Errorf("Invalid annotation value for key %q: %s", constants.SidecarInjectionAnnotation, inject)
 	}
 	return
-}
-
-func isAnnotatedForPortExclusion(annotations map[string]string, portAnnotation string, objectKind string, objectName string) (ports []int, err error) {
-	portsToExcludeStr, ok := annotations[portAnnotation]
-	if !ok {
-		return ports, err
-	}
-
-	log.Trace().Msgf("%s %s has port exclusion annotation: '%s:%s'", objectKind, objectName, portAnnotation, portsToExcludeStr)
-	portsToExclude := strings.Split(portsToExcludeStr, ",")
-	for _, portStr := range portsToExclude {
-		portStr := strings.TrimSpace(portStr)
-		portInt, ok := strconv.Atoi(portStr)
-		if ok != nil || portInt <= 0 {
-			err = errors.Errorf("Invalid port '%s' specified for annotation '%s'", portStr, portAnnotation)
-			ports = nil
-			return ports, err
-		}
-		ports = append(ports, portInt)
-	}
-	return ports, err
 }
 
 func patchAdmissionResponse(resp *admissionv1.AdmissionResponse, patchBytes []byte) {
