+++ pkg/injector/webhook_test.go
@@ -737,166 +737,6 @@ var _ = Describe("Testing Injector Functions", func() {
 	})
 })
 
-func TestIsAnnotatedForPortExclusion(t *testing.T) {
-	testCases := []struct {
-		name          string
-		annotations   map[string]string
-		forAnnotation string
-		expectedError error
-		expectedPorts []int
-	}{
-		{
-			name:          "contains outbound port exclusion list annotation",
-			annotations:   map[string]string{outboundPortExclusionListAnnotation: "6060, 7070"},
-			forAnnotation: outboundPortExclusionListAnnotation,
-			expectedError: nil,
-			expectedPorts: []int{6060, 7070},
-		},
-		{
-			name:          "contains inbound port exclusion list annotation",
-			annotations:   map[string]string{inboundPortExclusionListAnnotation: "6060, 7070"},
-			forAnnotation: inboundPortExclusionListAnnotation,
-			expectedError: nil,
-			expectedPorts: []int{6060, 7070},
-		},
-		{
-			name:          "does not contains port exclusion list annontation",
-			annotations:   nil,
-			forAnnotation: "",
-			expectedError: nil,
-			expectedPorts: nil,
-		},
-		{
-			name:          "contains outbound port exclusion list annontation but invalid port",
-			annotations:   map[string]string{outboundPortExclusionListAnnotation: "6060, -7070"},
-			forAnnotation: outboundPortExclusionListAnnotation,
-			expectedError: errors.Errorf("Invalid port '%s' specified for annotation '%s'", "-7070", outboundPortExclusionListAnnotation),
-			expectedPorts: nil,
-		},
-		{
-			name:          "contains inbound port exclusion list annontation but invalid port",
-			annotations:   map[string]string{inboundPortExclusionListAnnotation: "6060, -7070"},
-			forAnnotation: inboundPortExclusionListAnnotation,
-			expectedError: errors.Errorf("Invalid port '%s' specified for annotation '%s'", "-7070", inboundPortExclusionListAnnotation),
-			expectedPorts: nil,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			assert := tassert.New(t)
-
-			ports, err := isAnnotatedForPortExclusion(tc.annotations, tc.forAnnotation, "-kind-", "-name-")
-			if err != nil {
-				assert.EqualError(tc.expectedError, err.Error())
-			} else {
-				assert.Equal(tc.expectedError, err)
-			}
-			assert.ElementsMatch(tc.expectedPorts, ports)
-		})
-	}
-}
-
-func TestGetPodOutboundPortExclusionList(t *testing.T) {
-	testCases := []struct {
-		name          string
-		podAnnotation map[string]string
-		forAnnotation string
-		expectedError error
-		expectedPorts []int
-	}{
-		{
-			name:          "contains outbound port exclusion list annotation",
-			podAnnotation: map[string]string{outboundPortExclusionListAnnotation: "6060, 7070"},
-			forAnnotation: outboundPortExclusionListAnnotation,
-			expectedError: nil,
-			expectedPorts: []int{6060, 7070},
-		},
-		{
-			name:          "contains inbound port exclusion list annotation",
-			podAnnotation: map[string]string{inboundPortExclusionListAnnotation: "6060, 7070"},
-			forAnnotation: inboundPortExclusionListAnnotation,
-			expectedError: nil,
-			expectedPorts: []int{6060, 7070},
-		},
-		{
-			name:          "does not contains any port exclusion list annontation",
-			podAnnotation: nil,
-			forAnnotation: "",
-			expectedError: nil,
-			expectedPorts: nil,
-		},
-		{
-			name:          "contains outbound port exclusion list annontation but invalid port",
-			podAnnotation: map[string]string{outboundPortExclusionListAnnotation: "6060, -7070"},
-			forAnnotation: outboundPortExclusionListAnnotation,
-			expectedError: errors.Errorf("Invalid port '%s' specified for annotation '%s'", "-7070", outboundPortExclusionListAnnotation),
-			expectedPorts: nil,
-		},
-		{
-			name:          "contains inbound port exclusion list annontation but invalid port",
-			podAnnotation: map[string]string{inboundPortExclusionListAnnotation: "6060, -7070"},
-			forAnnotation: inboundPortExclusionListAnnotation,
-			expectedError: errors.Errorf("Invalid port '%s' specified for annotation '%s'", "-7070", inboundPortExclusionListAnnotation),
-			expectedPorts: nil,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			assert := tassert.New(t)
-
-			mockCtrl := gomock.NewController(GinkgoT())
-			mockKubeController := k8s.NewMockController(mockCtrl)
-			fakeClientSet := fake.NewSimpleClientset()
-			namespace := "test"
-			osmNamespace := "osm-namespace"
-
-			wh := &mutatingWebhook{
-				kubeClient:     fakeClientSet,
-				kubeController: mockKubeController,
-				osmNamespace:   osmNamespace,
-				nonInjectNamespaces: mapset.NewSetFromSlice([]interface{}{
-					metav1.NamespaceSystem,
-					metav1.NamespacePublic,
-					osmNamespace,
-				}),
-			}
-
-			testNamespace := &corev1.Namespace{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: namespace,
-				},
-			}
-			retNs, _ := fakeClientSet.CoreV1().Namespaces().Create(context.TODO(), testNamespace, metav1.CreateOptions{})
-
-			pod := &corev1.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:        "pod-test",
-					Annotations: tc.podAnnotation,
-				},
-				Spec: corev1.PodSpec{
-					ServiceAccountName: "test-SA",
-				},
-			}
-
-			_, err := fakeClientSet.CoreV1().Pods(namespace).Create(context.TODO(), pod, metav1.CreateOptions{})
-			Expect(err).ToNot(HaveOccurred())
-
-			mockKubeController.EXPECT().IsMonitoredNamespace(namespace).Return(true).Times(1)
-			mockKubeController.EXPECT().GetNamespace(namespace).Return(retNs)
-
-			ports, err := wh.getPortExclusionListForPod(pod, namespace, tc.forAnnotation)
-			if err != nil {
-				assert.EqualError(tc.expectedError, err.Error())
-			} else {
-				assert.Equal(tc.expectedError, err)
-			}
-			assert.ElementsMatch(tc.expectedPorts, ports)
-		})
-	}
-}
-
 func TestPodCreationHandler(t *testing.T) {
 	tests := []struct {
 		name                 string
